Combine period.py and plot_periodogram.py
