+++ gammapy/time/tests/test_period.py
@@ -3,7 +3,8 @@ import pytest
 import numpy as np
 from numpy.testing import assert_allclose
 from astropy.stats import LombScargle
-from ..period import robust_periodogram
+from ...utils.testing import requires_dependency
+from ..period import robust_periodogram, plot_periodogram
 
 pytest.importorskip("astropy", "3.0")
 
@@ -174,3 +175,51 @@ def test_period(pars):
     assert_allclose(fap["davies"], pars["fap"]["davies"], rtol=1e-3)
     assert_allclose(fap["baluev"], pars["fap"]["baluev"], rtol=1e-3)
     assert_allclose(fap["bootstrap"], pars["fap"]["bootstrap"], rtol=1e-3)
+
+
+@requires_dependency("matplotlib")
+def test_plot_periodogram():
+    pars = dict(
+        period=7,
+        amplitude=2,
+        t_length=100,
+        n_data=1000,
+        n_obs=500,
+        n_outliers=50,
+        loss="cauchy",
+        scale=1,
+    )
+    test_data = simulate_test_data(
+        pars["period"],
+        pars["amplitude"],
+        pars["t_length"],
+        pars["n_data"],
+        pars["n_obs"],
+        pars["n_outliers"],
+    )
+
+    periodogram = robust_periodogram(
+        test_data["t"],
+        test_data["y"],
+        test_data["dy"],
+        loss=pars["loss"],
+        scale=pars["scale"],
+    )
+
+    fap = fap_astropy(
+        periodogram["power"],
+        1.0 / periodogram["periods"],
+        test_data["t"],
+        test_data["y"],
+        test_data["dy"],
+    )
+
+    plot_periodogram(
+        test_data["t"],
+        test_data["y"],
+        periodogram["periods"],
+        periodogram["power"],
+        test_data["dy"],
+        periodogram["best_period"],
+        max(fap.values()),
+    )
